Add SharedMemory class implementation
